+++ src/markdown/utils.ts
@@ -4,7 +4,8 @@ import {
   CLOSING_STYLE_TOKENS,
   INLINE_ELEMENT_TOKENS,
   SIMPLE_MATH_DELIM_RE,
-  latexEnvironments
+  latexEnvironments,
+  attrsSharedMarker
 } from "./common/consts";
 
 export const endTag = (arg: string, shouldBeFirst = false): RegExp  => {
@@ -743,9 +744,6 @@ export const getSpacesFromLeft = (str: string) => {
   return str ? str.length - strTrimLeft.length : 0;
 };
 
-// Shared-attrs marker (see md-block-rule/begin-tabular/tabular-td.ts).
-const ATTRS_SHARED = Symbol.for('mathpix.tabular.attrsShared');
-
 /** add additional attributes to the parent token */
 export const addAttributesToParentTokenByType = (
   parentToken: Token,
@@ -763,7 +761,7 @@ export const addAttributesToParentTokenByType = (
     }
     if (token.children.find(item => item.type === tokenType)) {
       if (parentToken.attrs?.length) {
-        if ((parentToken.attrs as any)[ATTRS_SHARED]) {
+        if ((parentToken.attrs as any)[attrsSharedMarker]) {
           parentToken.attrs = parentToken.attrs.map(p => [p[0], p[1]]);
         }
         parentToken.attrs.push(attrs)